chore(ci): Rename ZT-bnt-common to bnt-common for gitlab (#4)
Summary
This PR changes the name of the ZT-bnt-common repo reference to bnt-common for use in gitlab. This is an effort to create consistent naming in gitlab and simplify git-sync-api logic introduced through the github migration. Functionally this should not result in any changes to CI.

Test Plan
Verify certificate-cap-test-plan job is still picked up and ran upon a CI run.

Ticket
ZTBNT-2080